running seperated conode-library
